worker/requirements.txt
pyspark==3.5.4
pymongo==4.10.1
pandas==2.2.3
textblob==0.18.0
flask==3.1.0
python-dotenv==1.0.1
